Refactor authentication middleware and validation
- Move registration and login validation middleware to separate files
- Update import paths for firebase connection
- Update middleware function names for clarity
- Remove unused express-validator import in authRoute.js
- Update eventoRoute.js to import validateLogin and validateRegister from authMiddleware.js
