+++ backend/controllers/authController.js
@@ -1,6 +1,6 @@
 require('dotenv').config();
 const bcrypt = require('bcrypt');
-const { firebaseApp } = require('../firebase/firebaseConnection');
+const { firebaseApp } = require('../db/firebaseConnection');
 const { getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword } = require('firebase/auth');
 const { body, validationResult } = require('express-validator');
 const jwt = require('jsonwebtoken');
@@ -32,20 +32,6 @@ const getFirebaseErrorMessage = (errorCode) => {
   }
 };
 
-// TODO: Move to middleware folder's
-// Middleware for registration validation
-const validateRegister = [
-  body('nombre').notEmpty().withMessage('Nombre is required.'),
-  body('email').isEmail().withMessage('Please provide a valid email address.'),
-  body('password').isLength({ min: 6 }).withMessage('Password must be at least 6 characters long.'),
-];
-
-// Middleware for login validation
-const validateLogin = [
-  body('email').isEmail().withMessage('Please provide a valid email address.'),
-  body('password').notEmpty().withMessage('Password cannot be empty.'),
-];
-
 // FIREBASE REGISTER AND CREATE USER IN MONGODB
 const registerEmailPassword = async (req, res) => {
   const errors = validationResult(req);
@@ -54,31 +40,35 @@ const registerEmailPassword = async (req, res) => {
   }
 
   const { nombre, username, email, password, rol = 'Participante' } = req.body; // Se puede definir un rol por defecto
+  const hashedPassword = await bcrypt.hash(password, 10); // Hashear la contraseña antes de registrarla
+  const token = generateJWT(userCredential.user.uid); // Generate JWT
+
+  const newUser = new User({
+    nombre,
+    username,
+    email,
+    rol,
+    hashedPassword, // Almacenar la contraseña hasheada
+    firebaseId: userCredential.user.uid // Guardar el uid de Firebase
+  });
 
   try {
-    // Hashear la contraseña antes de registrarla
-    const hashedPassword = await bcrypt.hash(password, 10);
-
-    // Registrar el usuario en Firebase
-    const userCredential = await createUserWithEmailAndPassword(auth, email, password);
-    const token = generateJWT(userCredential.user.uid); // Generate JWT
-
     // Crear el usuario en MongoDB con la contraseña hasheada
-    const newUser = new User({
-      nombre,
-      username,
-      email,
-      rol,
-      hashedPassword, // Almacenar la contraseña hasheada
-      firebaseId: userCredential.user.uid // Guardar el uid de Firebase
-    });
-
     await newUser.save();
 
+    try {
+      // Registrar el usuario en Firebase
+      const userCredential = await createUserWithEmailAndPassword(auth, email, password);
+
+    }catch(error){
+      const { message, status } = getFirebaseErrorMessage(error.code);
+      return res.status(status || 500).json({ message: message || error.message, errorCode: error.code });
+    }
+    // Retornar el token y el usuario registrado
     return res.status(201).json({ message: 'User registered successfully.', token, user: newUser });
+
   } catch (error) {
-    const { message, status } = getFirebaseErrorMessage(error.code);
-    return res.status(status || 500).json({ message: message || error.message, errorCode: error.code });
+      return res.status(500).json({ message: 'Error registering user.', error: error.message });
   }
 };
 
@@ -112,4 +102,4 @@ const loginEmailPassword = async (req, res) => {
   }
 };
 
-module.exports = { registerEmailPassword, loginEmailPassword, validateRegister, validateLogin };
+module.exports = { registerEmailPassword, loginEmailPassword };
