+++ backend/db/dbConnection.js
@@ -8,9 +8,9 @@ const dbConnection = async () => {
 	try {
     //mongoose connection events, async functions to validate status
     mongoose.connection.on('disconnected', () => console.log('Disconnected from mongoDB instance'));
-    mongoose.connection.on('connected', () => console.log('Connected to mongoDB instance'));
+    mongoose.connection.on('connected', () => console.log(`Connected to mongoDB instance: ${MONGO_URI}`));
     mongoose.connection.on('close', () => console.log('Connection closed by peer'));
-    console.log(`Connecting to mongoDB instance @ ${MONGO_URI}`);
+    console.log(`Establishing connection to mongo cluster...`);
 		mongoose.connect(ATLAS_URI, {
 			autoIndex: true,
 		});
