+++ backend/db/firebaseConnection.js
